taskmanager: fold event helpers into TaskHandle
TaskHandle is the one user-facing style: the former TaskHandler verbs
over the event channel, synchronous bodies included. The package-level
event constructors (NewStatusUpdate/Working/Completed/NewArtifactUpdate
and the Events channel helper) are gone - they were a third vocabulary
sitting between the handle and the raw contract. ReplyText stays as the
message builder both styles use. The README teaches the TaskHandle
style; examples/simple-v2 keeps demonstrating the raw channel contract
with hand-built events.
